backend/requirements.txt
Flask>=3.1.0
flask-cors>=5.0.1 
python-dotenv>=1.1.0
google-api-python-client>=2.165.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
requests>=2.32.3
uuid